#82 refactor: create deep issue module with hidden label serialization
- Create src/lib/modules/issues/ merging context + commands + types
- Hide label JSON serialization/deserialization inside module
- Narrow generated Issue type with proper literal unions for priority/status/worktree_state
- Export public types: Issue, IssueLabel, SortMode, CreateIssueRequest, UpdateIssueRequest
- Delete issues.context.svelte.ts, issue_commands.ts, types/issue.ts
