+++ src/lib/components/DashboardToolbar.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { RefreshCw, Scissors } from 'lucide-svelte';
-	import type { SortMode } from '$lib/context/issues.context.svelte.js';
+	import type { SortMode } from '$lib/modules/issues/index.svelte.js';
 	import type { ViewMode } from '$lib/context/view_preference.context.svelte.js';
 	import ViewToggle from '$lib/components/ViewToggle.svelte';
