+++ src/lib/components/IssueEditDialog.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { Issue, UpdateIssueRequest } from '$lib/types/issue';
+	import type { Issue, UpdateIssueRequest } from '$lib/modules/issues/index.svelte.js';
 	import { FALLBACK_ISSUE_COLOR } from '$lib/types/color_palette';
 	import PaletteColorPicker from './PaletteColorPicker.svelte';
