+++ src/lib/engine/map_issue_to_state_dimensions.ts
@@ -1,4 +1,4 @@
-import type { Issue } from '$lib/types/issue';
+import type { Issue } from '$lib/modules/issues/index.svelte.js';
 import type { GitStatusCache, ExecutionPhase, SessionState } from '$lib/types/generated';
 import type {
 	ForestBranchStatus,
