--- src/lib/types/issue.ts
@@ -1,57 +1,0 @@
-import type { WorktreeState } from './worktree';
-
-export interface IssueLabel {
-	name: string;
-	color: string;
-}
-
-export interface Issue {
-	id: string;
-	dashboard_id: string;
-	name: string;
-	priority: 'low' | 'medium' | 'high' | 'top' | null;
-	color: string | null;
-	status: 'active' | 'archived';
-	github_issue_url: string | null;
-	github_issue_number: number | null;
-	branch_name: string | null;
-	base_branch: string | null;
-	worktree_folder: string | null;
-	worktree_state: WorktreeState;
-	parent_issue_id: string | null;
-	editor_folder: string | null;
-	dev_server_command: string | null;
-	dev_server_port: number | null;
-	dev_server_pid: number | null;
-	browser_url: string | null;
-	labels: IssueLabel[];
-	sort_order: number;
-	created_at: string;
-}
-
-export interface CreateIssueRequest {
-	dashboard_id: string;
-	name: string;
-	priority?: 'low' | 'medium' | 'high' | 'top' | null;
-	color?: string | null;
-	github_issue_url?: string | null;
-	github_issue_number?: number | null;
-	parent_issue_id?: string | null;
-	labels?: string | null;
-}
-
-export interface UpdateIssueRequest {
-	id: string;
-	name?: string;
-	priority?: 'low' | 'medium' | 'high' | 'top' | null;
-	color?: string | null;
-	github_issue_url?: string | null;
-	github_issue_number?: number | null;
-	branch_name?: string | null;
-	base_branch?: string | null;
-	worktree_folder?: string | null;
-	worktree_state?: WorktreeState;
-	parent_issue_id?: string | null;
-	labels?: string | null;
-	sort_order?: number;
-}
